QuantumGraphObject class is now able to update the scattering and transmission matrices based on the structure of the graph itself and the scattering matrices at each node. This saves an awful lot of time.
The code has been tested on the Neumann tetrahedron, and gives the same result as a hand-written scattering matrix does.

Currently uncommented. In fact, code is currently very hacked together and must be cleaned up and have far more documentation due to the complexity, but it works in this state.