bound_qg_roots now takes in an input stream that is created by one of the graph definitions (haven't added graph_defs/ to the git archive yet). This file contains one of the QuantumGraph header strings that can initialize the scattering and transmission matrices of the graph.
find_bounded_qg_roots now reads its quantum graph directly from the header at the top of the output file from bound_qg_roots. This allows us to send the same problem along to programs that might run seperately.

quantumgraphrootfinding.cpp is still unfinished, but the changes made here were for purposes of testing. Nothing important was changed.

The makefile now creates the data file for the root boundaries instead of piping the output to the root-finding program directly. This is mostly to make sure that I have a record of the bounds when I want to check them. There is still an issue with missing roots when the roots are very close together.